requirements.txt
flask
flask-sqlalchemy
flask-login
python-dotenv
werkzeug
psycopg2-binary
gunicorn